feat: #6 rename `mover.py` to `filemover.py`
closes #6
