+++ tests/test_mover.py
@@ -1,6 +1,6 @@
 import unittest
-from src.mover import Mover
 from src.mover_config import MoverConfig
+from src.filemover import Mover
 
 class DummyRenameConfig:
     def apply_rename(self, name):
